common/cli/dt-core/pyproject.toml
[project]
name = "dt-core"
version = "0.1.0"
description = "Unified CLI for development tools"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.scripts]
dt = "dt_core.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dt_core"]
